backend/requirement.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.3
django-allauth==0.42.0
django-bootstrap3==11.0.0
django-braces==1.13.0
django-cors-headers==3.5.0
django-humanize==0.1.2
django-misaka==0.2.1
django-rest-auth==0.9.5
djangorestframework==3.11.1
houdini.py==0.1.0
humanize==0.5.1
idna==2.10
misaka==2.1.1
oauthlib==3.1.0
Pillow==7.2.0
psycopg2==2.8.5
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
